+++ packages/fs-detectors/src/detect-builders.ts
@@ -8,10 +8,10 @@ import type {
   Builder,
   Config,
   BuilderFunctions,
-  Services,
   ExperimentalServices,
   ProjectSettings,
   Service,
+  ExperimentalServicesV2,
 } from '@vercel/build-utils';
 import { isOfficialRuntime } from './is-official-runtime';
 import {
@@ -65,8 +65,8 @@ export interface ErrorResponse {
 export interface Options {
   tag?: string;
   functions?: BuilderFunctions;
-  services?: Services;
   experimentalServices?: ExperimentalServices;
+  experimentalServicesV2?: ExperimentalServicesV2;
   ignoreBuildScript?: boolean;
   projectSettings?: ProjectSettings;
   cleanUrls?: boolean;
@@ -143,11 +143,16 @@ export async function detectBuilders(
   services?: Service[];
   useImplicitEnvInjection?: boolean;
 }> {
-  const { services, experimentalServices, projectSettings = {} } = options;
+  const {
+    experimentalServices,
+    experimentalServicesV2,
+    projectSettings = {},
+  } = options;
   const { framework } = projectSettings;
-  const configuredServices = services ?? experimentalServices;
-  const configuredServicesType =
-    services != null ? 'services' : 'experimentalServices';
+  const configuredServices = experimentalServices ?? experimentalServicesV2;
+  const configuredServicesType = experimentalServices
+    ? 'experimentalServices'
+    : 'experimentalServicesV2';
   const hasServicesConfig =
     configuredServices != null && typeof configuredServices === 'object';