+++ packages/build-utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @vercel/build-utils
 
+## 13.27.0
+
+### Minor Changes
+
+- 338cc35: Add isPackageInstalled util for detecting dependencies during build.
+  Fix Vercel Flags dependency detection for emitting datafiles during builds with OIDC tokens.
+
+## 13.26.6
+
+### Patch Changes
+
+- 3019788: [services] Remove the `services` field from `vercel.json` and the `VERCEL_USE_SERVICES` gate.
+- fe893ec: [services] Add `experimentalServicesV2` field to `vercel.json` implementing the new schema for services.
+
+## 13.26.5
+
+### Patch Changes
+
+- 1180675: Revert "[flags] fix dep detection for build embedding (#16242)"
+
 ## 13.26.4
 
 ### Patch Changes