+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vercel/elysia
 
+## 0.1.87
+
+### Patch Changes
+
+- @vercel/node@5.8.11
+
+## 0.1.86
+
+### Patch Changes
+
+- @vercel/node@5.8.10
+
+## 0.1.85
+
+### Patch Changes
+
+- @vercel/node@5.8.9
+
 ## 0.1.84
 
 ### Patch Changes